Move some string literal unions to generated file
