+++ sdk/search/search-documents/src/index.ts
@@ -183,6 +183,33 @@ export {
   VectorSearchProfile,
   WordDelimiterTokenFilter,
 } from "./generated/service/models";
+export {
+  BlobIndexerDataToExtract,
+  BlobIndexerImageAction,
+  BlobIndexerParsingMode,
+  BlobIndexerPDFTextRotationAlgorithm,
+  CustomEntityLookupSkillLanguage,
+  EntityCategory,
+  EntityRecognitionSkillLanguage,
+  ImageAnalysisSkillLanguage,
+  ImageDetail,
+  IndexerExecutionEnvironment,
+  KeyPhraseExtractionSkillLanguage,
+  OcrSkillLanguage,
+  PIIDetectionSkillMaskingMode,
+  RegexFlags,
+  SearchIndexerDataSourceType,
+  SentimentSkillLanguage,
+  SplitSkillLanguage,
+  TextSplitMode,
+  TextTranslationSkillLanguage,
+  VectorFilterMode,
+  VectorQueryKind,
+  VectorSearchAlgorithmKind,
+  VectorSearchAlgorithmMetric,
+  VectorSearchVectorizerKind,
+  VisualFeature,
+} from "./generatedStringLiteralUnions";
 export { default as GeographyPoint } from "./geographyPoint";
 export { IndexDocumentsBatch } from "./indexDocumentsBatch";
 export {
@@ -236,11 +263,9 @@ export {
   SuggestResult,
   UnionToIntersection,
   UploadDocumentsOptions,
-  VectorFilterMode,
   VectorizableTextQuery,
   VectorizedQuery,
   VectorQuery,
-  VectorQueryKind,
   VectorSearchOptions,
 } from "./indexModels";
 export { odata } from "./odata";
@@ -264,10 +289,6 @@ export {
   AzureOpenAIVectorizer,
   BaseVectorSearchAlgorithmConfiguration,
   BaseVectorSearchVectorizer,
-  BlobIndexerDataToExtract,
-  BlobIndexerImageAction,
-  BlobIndexerParsingMode,
-  BlobIndexerPDFTextRotationAlgorithm,
   CharFilter,
   CognitiveServicesAccount,
   ComplexDataType,
@@ -286,7 +307,6 @@ export {
   CreateSynonymMapOptions,
   CustomAnalyzer,
   CustomEntityLookupSkill,
-  CustomEntityLookupSkillLanguage,
   CustomVectorizer,
   CustomVectorizerParameters,
   DataChangeDetectionPolicy,
@@ -298,9 +318,7 @@ export {
   DeleteSkillsetOptions,
   DeleteSynonymMapOptions,
   EdgeNGramTokenFilter,
-  EntityCategory,
   EntityRecognitionSkill,
-  EntityRecognitionSkillLanguage,
   ExhaustiveKnnAlgorithmConfiguration,
   ExhaustiveKnnParameters,
   GetAliasOptions,
@@ -315,16 +333,12 @@ export {
   HnswAlgorithmConfiguration,
   HnswParameters,
   ImageAnalysisSkill,
-  ImageAnalysisSkillLanguage,
-  ImageDetail,
-  IndexerExecutionEnvironment,
   IndexingParameters,
   IndexingParametersConfiguration,
   IndexIterator,
   IndexNameIterator,
   IndexProjectionMode,
   KeyPhraseExtractionSkill,
-  KeyPhraseExtractionSkillLanguage,
   KeywordTokenizer,
   KnownAnalyzerNames,
   KnownCharFilterNames,
@@ -342,12 +356,9 @@ export {
   LuceneStandardTokenizer,
   NGramTokenFilter,
   OcrSkill,
-  OcrSkillLanguage,
   PatternAnalyzer,
   PatternTokenizer,
   PIIDetectionSkill,
-  PIIDetectionSkillMaskingMode,
-  RegexFlags,
   ResetDocumentsOptions,
   ResetIndexerOptions,
   ResetSkillsOptions,
@@ -362,7 +373,6 @@ export {
   SearchIndexerCache,
   SearchIndexerDataIdentity,
   SearchIndexerDataSourceConnection,
-  SearchIndexerDataSourceType,
   SearchIndexerIndexProjections,
   SearchIndexerIndexProjectionsParameters,
   SearchIndexerKnowledgeStore,
@@ -373,23 +383,15 @@ export {
   SearchResourceEncryptionKey,
   SearchServiceStatistics,
   SentimentSkill,
-  SentimentSkillLanguage,
   SimilarityAlgorithm,
   SimpleField,
   SplitSkill,
-  SplitSkillLanguage,
   SynonymMap,
-  TextSplitMode,
   TextTranslationSkill,
-  TextTranslationSkillLanguage,
   TokenFilter,
   VectorSearch,
   VectorSearchAlgorithmConfiguration,
-  VectorSearchAlgorithmKind,
-  VectorSearchAlgorithmMetric,
   VectorSearchVectorizer,
-  VectorSearchVectorizerKind,
-  VisualFeature,
   WebApiSkill,
 } from "./serviceModels";
 export { createSynonymMapFromFile } from "./synonymMapHelper";
