+++ sdk/search/search-documents/src/serviceModels.ts
@@ -82,6 +82,31 @@ import {
   VectorSearchProfile,
   WordDelimiterTokenFilter,
 } from "./generated/service/models";
+import {
+  BlobIndexerDataToExtract,
+  BlobIndexerImageAction,
+  BlobIndexerParsingMode,
+  BlobIndexerPDFTextRotationAlgorithm,
+  CustomEntityLookupSkillLanguage,
+  EntityCategory,
+  EntityRecognitionSkillLanguage,
+  ImageAnalysisSkillLanguage,
+  ImageDetail,
+  IndexerExecutionEnvironment,
+  KeyPhraseExtractionSkillLanguage,
+  OcrSkillLanguage,
+  PIIDetectionSkillMaskingMode,
+  RegexFlags,
+  SearchIndexerDataSourceType,
+  SentimentSkillLanguage,
+  SplitSkillLanguage,
+  TextSplitMode,
+  TextTranslationSkillLanguage,
+  VectorSearchAlgorithmKind,
+  VectorSearchAlgorithmMetric,
+  VectorSearchVectorizerKind,
+  VisualFeature,
+} from "./generatedStringLiteralUnions";
 
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
 
@@ -2254,17 +2279,10 @@ export interface SearchIndexerKnowledgeStoreParameters {
   synthesizeGeneratedKeyName?: boolean;
 }
 
-/** The similarity metric to use for vector comparisons. */
-export type VectorSearchAlgorithmMetric = "cosine" | "euclidean" | "dotProduct";
-
-export type VectorSearchAlgorithmKind = "hnsw" | "exhaustiveKnn";
-
 /**
  * Defines behavior of the index projections in relation to the rest of the indexer.
  */
 export type IndexProjectionMode = "skipIndexingParentDocuments" | "includeIndexingParentDocuments";
-
-export type VectorSearchVectorizerKind = "azureOpenAI" | "customWebApi";
 
 /** A dictionary of indexer-specific configuration properties. Each name is the name of a specific property. Each value must be of a primitive type. */
 export interface IndexingParametersConfiguration {
@@ -2453,549 +2471,5 @@ export interface ImageAnalysisSkill extends BaseSearchIndexerSkill {
   /** A string indicating which domain-specific details to return. */
   details?: ImageDetail[];
 }
-
-/**
- * ### Known values supported by the service
- * **storageMetadata**: Indexes just the standard blob properties and user-specified metadata.
- * **allMetadata**: Extracts metadata provided by the Azure blob storage subsystem and the content-type specific metadata (for example, metadata unique to just .png files are indexed).
- * **contentAndMetadata**: Extracts all metadata and textual content from each blob.
- */
-export type BlobIndexerDataToExtract = "storageMetadata" | "allMetadata" | "contentAndMetadata";
-/**
- * ### Known values supported by the service
- * **none**: Ignores embedded images or image files in the data set.  This is the default.
- * **generateNormalizedImages**: Extracts text from images (for example, the word "STOP" from a traffic stop sign), and embeds it into the content field.  This action requires that "dataToExtract" is set to "contentAndMetadata".  A normalized image refers to additional processing resulting in uniform image output, sized and rotated to promote consistent rendering when you include images in visual search results. This information is generated for each image when you use this option.
- * **generateNormalizedImagePerPage**: Extracts text from images (for example, the word "STOP" from a traffic stop sign), and embeds it into the content field, but treats PDF files differently in that each page will be rendered as an image and normalized accordingly, instead of extracting embedded images.  Non-PDF file types will be treated the same as if "generateNormalizedImages" was set.
- */
-export type BlobIndexerImageAction =
-  | "none"
-  | "generateNormalizedImages"
-  | "generateNormalizedImagePerPage";
-/**
- * ### Known values supported by the service
- * **none**: Leverages normal text extraction.  This is the default.
- * **detectAngles**: May produce better and more readable text extraction from PDF files that have rotated text within them.  Note that there may be a small performance speed impact when this parameter is used.  This parameter only applies to PDF files, and only to PDFs with embedded text.  If the rotated text appears within an embedded image in the PDF, this parameter does not apply.
- */
-export type BlobIndexerPDFTextRotationAlgorithm = "none" | "detectAngles";
-/**
- * ### Known values supported by the service
- * **default**: Set to default for normal file processing.
- * **text**: Set to text to improve indexing performance on plain text files in blob storage.
- * **delimitedText**: Set to delimitedText when blobs are plain CSV files.
- * **json**: Set to json to extract structured content from JSON files.
- * **jsonArray**: Set to jsonArray to extract individual elements of a JSON array as separate documents in Azure Cognitive Search.
- * **jsonLines**: Set to jsonLines to extract individual JSON entities, separated by a new line, as separate documents in Azure Cognitive Search.
- */
-export type BlobIndexerParsingMode =
-  | "default"
-  | "text"
-  | "delimitedText"
-  | "json"
-  | "jsonArray"
-  | "jsonLines";
-/**
- * Defines supported languages for {@link CustomEntityLookupSkill}
- * {@link KnownCustomEntityLookupSkillLanguage} can be used interchangeably with this type
- */
-export type CustomEntityLookupSkillLanguage =
-  | "da"
-  | "de"
-  | "en"
-  | "es"
-  | "fi"
-  | "fr"
-  | "it"
-  | "ko"
-  | "pt";
-
-export type EntityCategory =
-  | "location"
-  | "organization"
-  | "person"
-  | "quantity"
-  | "datetime"
-  | "url"
-  | "email";
-
-/**
- * Defines supported languages for {@link EntityRecognitionSkill}
- * {@link KnownEntityRecognitionSkillLanguage} can be used interchangeably with this type
- */
-export type EntityRecognitionSkillLanguage =
-  | "ar"
-  | "cs"
-  | "zh-Hans"
-  | "zh-Hant"
-  | "da"
-  | "nl"
-  | "en"
-  | "fi"
-  | "fr"
-  | "de"
-  | "el"
-  | "hu"
-  | "it"
-  | "ja"
-  | "ko"
-  | "no"
-  | "pl"
-  | "pt-PT"
-  | "pt-BR"
-  | "ru"
-  | "es"
-  | "sv"
-  | "tr";
-
-/**
- * Defines supported languages for {@link ImageAnalysisSkill}
- * {@link KnownImageAnalysisSkillLanguage} can be used interchangeably with this type
- */
-export type ImageAnalysisSkillLanguage =
-  | "ar"
-  | "az"
-  | "bg"
-  | "bs"
-  | "ca"
-  | "cs"
-  | "cy"
-  | "da"
-  | "de"
-  | "el"
-  | "en"
-  | "es"
-  | "et"
-  | "eu"
-  | "fi"
-  | "fr"
-  | "ga"
-  | "gl"
-  | "he"
-  | "hi"
-  | "hr"
-  | "hu"
-  | "id"
-  | "it"
-  | "ja"
-  | "kk"
-  | "ko"
-  | "lt"
-  | "lv"
-  | "mk"
-  | "ms"
-  | "nb"
-  | "nl"
-  | "pl"
-  | "prs"
-  | "pt-BR"
-  | "pt"
-  | "pt-PT"
-  | "ro"
-  | "ru"
-  | "sk"
-  | "sl"
-  | "sr-Cyrl"
-  | "sr-Latn"
-  | "sv"
-  | "th"
-  | "tr"
-  | "uk"
-  | "vi"
-  | "zh"
-  | "zh-Hans"
-  | "zh-Hant";
-
-export type ImageDetail = "celebrities" | "landmarks";
-
-/**
- * ### Known values supported by the service
- * **standard**: Indicates that Azure Cognitive Search can determine where the indexer should execute. This is the default environment when nothing is specified and is the recommended value.
- * **private**: Indicates that the indexer should run with the environment provisioned specifically for the search service. This should only be specified as the execution environment if the indexer needs to access resources securely over shared private link resources.
- */
-export type IndexerExecutionEnvironment = "standard" | "private";
-
-/**
- * Defines supported languages for {@link KeyPhraseExtractionSkill}
- * {@link KnownKeyPhraseExtractionSkillLanguage} can be used interchangeably with this type
- */
-export type KeyPhraseExtractionSkillLanguage =
-  | "da"
-  | "nl"
-  | "en"
-  | "fi"
-  | "fr"
-  | "de"
-  | "it"
-  | "ja"
-  | "ko"
-  | "no"
-  | "pl"
-  | "pt-PT"
-  | "pt-BR"
-  | "ru"
-  | "es"
-  | "sv";
-
-/**
- * Defines supported languages for {@link  OcrSkill}
- * {@link KnownOcrSkillLanguage} can be used interchangeably with this type
- */
-export type OcrSkillLanguage =
-  | "af"
-  | "sq"
-  | "anp"
-  | "ar"
-  | "ast"
-  | "awa"
-  | "az"
-  | "bfy"
-  | "eu"
-  | "be"
-  | "be-cyrl"
-  | "be-latn"
-  | "bho"
-  | "bi"
-  | "brx"
-  | "bs"
-  | "bra"
-  | "br"
-  | "bg"
-  | "bns"
-  | "bua"
-  | "ca"
-  | "ceb"
-  | "rab"
-  | "ch"
-  | "hne"
-  | "zh-Hans"
-  | "zh-Hant"
-  | "kw"
-  | "co"
-  | "crh"
-  | "hr"
-  | "cs"
-  | "da"
-  | "prs"
-  | "dhi"
-  | "doi"
-  | "nl"
-  | "en"
-  | "myv"
-  | "et"
-  | "fo"
-  | "fj"
-  | "fil"
-  | "fi"
-  | "fr"
-  | "fur"
-  | "gag"
-  | "gl"
-  | "de"
-  | "gil"
-  | "gon"
-  | "el"
-  | "kl"
-  | "gvr"
-  | "ht"
-  | "hlb"
-  | "hni"
-  | "bgc"
-  | "haw"
-  | "hi"
-  | "mww"
-  | "hoc"
-  | "hu"
-  | "is"
-  | "smn"
-  | "id"
-  | "ia"
-  | "iu"
-  | "ga"
-  | "it"
-  | "ja"
-  | "Jns"
-  | "jv"
-  | "kea"
-  | "kac"
-  | "xnr"
-  | "krc"
-  | "kaa-cyrl"
-  | "kaa"
-  | "csb"
-  | "kk-cyrl"
-  | "kk-latn"
-  | "klr"
-  | "kha"
-  | "quc"
-  | "ko"
-  | "kfq"
-  | "kpy"
-  | "kos"
-  | "kum"
-  | "ku-arab"
-  | "ku-latn"
-  | "kru"
-  | "ky"
-  | "lkt"
-  | "la"
-  | "lt"
-  | "dsb"
-  | "smj"
-  | "lb"
-  | "bfz"
-  | "ms"
-  | "mt"
-  | "kmj"
-  | "gv"
-  | "mi"
-  | "mr"
-  | "mn"
-  | "cnr-cyrl"
-  | "cnr-latn"
-  | "nap"
-  | "ne"
-  | "niu"
-  | "nog"
-  | "sme"
-  | "nb"
-  | "no"
-  | "oc"
-  | "os"
-  | "ps"
-  | "fa"
-  | "pl"
-  | "pt"
-  | "pa"
-  | "ksh"
-  | "ro"
-  | "rm"
-  | "ru"
-  | "sck"
-  | "sm"
-  | "sa"
-  | "sat"
-  | "sco"
-  | "gd"
-  | "sr"
-  | "sr-Cyrl"
-  | "sr-Latn"
-  | "xsr"
-  | "srx"
-  | "sms"
-  | "sk"
-  | "sl"
-  | "so"
-  | "sma"
-  | "es"
-  | "sw"
-  | "sv"
-  | "tg"
-  | "tt"
-  | "tet"
-  | "thf"
-  | "to"
-  | "tr"
-  | "tk"
-  | "tyv"
-  | "hsb"
-  | "ur"
-  | "ug"
-  | "uz-arab"
-  | "uz-cyrl"
-  | "uz"
-  | "vo"
-  | "wae"
-  | "cy"
-  | "fy"
-  | "yua"
-  | "za"
-  | "zu"
-  | "unk";
-
-/**
- * Defines values for PIIDetectionSkillMaskingMode.
- * ### Known values supported by the service
- * **none**: No masking occurs and the maskedText output will not be returned.
- * **replace**: Replaces the detected entities with the character given in the maskingCharacter parameter. The character will be repeated to the length of the detected entity so that the offsets will correctly correspond to both the input text as well as the output maskedText.
- */
-export type PIIDetectionSkillMaskingMode = "none" | "replace";
-
-/**
- * Defines flags for regex pattern matching
- * ### Known values supported by the service
- * **CANON_EQ**: Enables canonical equivalence.
- * **CASE_INSENSITIVE**: Enables case-insensitive matching.
- * **COMMENTS**: Permits whitespace and comments in the pattern.
- * **DOTALL**: Enables dotall mode.
- * **LITERAL**: Enables literal parsing of the pattern.
- * **MULTILINE**: Enables multiline mode.
- * **UNICODE_CASE**: Enables Unicode-aware case folding.
- * **UNIX_LINES**: Enables Unix lines mode.
- */
-export type RegexFlags =
-  | "CANON_EQ"
-  | "CASE_INSENSITIVE"
-  | "COMMENTS"
-  | "DOTALL"
-  | "LITERAL"
-  | "MULTILINE"
-  | "UNICODE_CASE"
-  | "UNIX_LINES";
-
-export type SearchIndexerDataSourceType =
-  | "azuresql"
-  | "cosmosdb"
-  | "azureblob"
-  | "azuretable"
-  | "mysql"
-  | "adlsgen2";
-
-/**
- * Defines supported languages for {@link SentimentSkill}
- * {@link KnownSentimentSkillLanguage} can be used interchangeably with this type
- */
-export type SentimentSkillLanguage =
-  | "da"
-  | "nl"
-  | "en"
-  | "fi"
-  | "fr"
-  | "de"
-  | "el"
-  | "it"
-  | "no"
-  | "pl"
-  | "pt-PT"
-  | "ru"
-  | "es"
-  | "sv"
-  | "tr";
-
-/**
- * Defines supported languages for {@link SplitSkill}
- * {@link KnownSplitSkillLanguage} can be used interchangeably with this type
- */
-export type SplitSkillLanguage =
-  | "am"
-  | "bs"
-  | "cs"
-  | "da"
-  | "de"
-  | "en"
-  | "es"
-  | "et"
-  | "fi"
-  | "fr"
-  | "he"
-  | "hi"
-  | "hr"
-  | "hu"
-  | "id"
-  | "is"
-  | "it"
-  | "ja"
-  | "ko"
-  | "lv"
-  | "nb"
-  | "nl"
-  | "pl"
-  | "pt"
-  | "pt-br"
-  | "ru"
-  | "sk"
-  | "sl"
-  | "sr"
-  | "sv"
-  | "tr"
-  | "ur"
-  | "zh";
-
-export type TextSplitMode = "pages" | "sentences";
-
-/**
- * Defines supported languages for {@link TextTranslationSkill}
- * {@link KnownTextTranslationSkillLanguage} can be used interchangeably with this type
- */
-export type TextTranslationSkillLanguage =
-  | "af"
-  | "ar"
-  | "bn"
-  | "bs"
-  | "bg"
-  | "yue"
-  | "ca"
-  | "zh-Hans"
-  | "zh-Hant"
-  | "hr"
-  | "cs"
-  | "da"
-  | "nl"
-  | "en"
-  | "et"
-  | "fj"
-  | "fil"
-  | "fi"
-  | "fr"
-  | "de"
-  | "el"
-  | "ht"
-  | "he"
-  | "hi"
-  | "mww"
-  | "hu"
-  | "is"
-  | "id"
-  | "it"
-  | "ja"
-  | "sw"
-  | "tlh"
-  | "tlh-Latn"
-  | "tlh-Piqd"
-  | "ko"
-  | "lv"
-  | "lt"
-  | "mg"
-  | "ms"
-  | "mt"
-  | "nb"
-  | "fa"
-  | "pl"
-  | "pt"
-  | "pt-br"
-  | "pt-PT"
-  | "otq"
-  | "ro"
-  | "ru"
-  | "sm"
-  | "sr-Cyrl"
-  | "sr-Latn"
-  | "sk"
-  | "sl"
-  | "es"
-  | "sv"
-  | "ty"
-  | "ta"
-  | "te"
-  | "th"
-  | "to"
-  | "tr"
-  | "uk"
-  | "ur"
-  | "vi"
-  | "cy"
-  | "yua"
-  | "ga"
-  | "kn"
-  | "mi"
-  | "ml"
-  | "pa";
-
-export type VisualFeature =
-  | "adult"
-  | "brands"
-  | "categories"
-  | "description"
-  | "faces"
-  | "objects"
-  | "tags";
 
 // END manually modified generated interfaces
